requirements.txt
# Core Data Science & ML
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
scipy>=1.7.0

# Visualization
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Web Framework
streamlit>=1.20.0
fastapi>=0.95.0
uvicorn>=0.20.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
networkx>=2.6.0

# Research & Analytics
lifelines>=0.27.0

# Additional dependencies for advanced features
requests>=2.28.0
python-multipart>=0.0.5
pydantic>=1.9.0 